requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
sentence-transformers==2.2.2
plotly==5.18.0
scikit-learn==1.3.2
python-Levenshtein==0.26.1
fuzzywuzzy==0.18.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.0.1
